applications/infrared-detect/requirements.txt
asyncio-nats-client
protobuf
